Move entities into this.state.entities instead of being top-level state
Allows an array to be used for the entities, which in turn allows
entities to be completely removed from state.

Previously, it was only possible to use an object because React does not
allow arrays as top-level state, which meant there was no way to delete
an entity due to the shallow-merge performed by React's `setState`;
omitting the entity to remove would not remove it, and setting it to
null would leave the key present, making the engine attempt to process
the null entity on the next update.